+++ server/Logging.h
@@ -2,13 +2,14 @@
 #define __LOGGING__
 
 #include <unistd.h>
+#include <stdint.h>
 #include "LogReport.h"
 
 void saveLogReport(LogReport message);
 void printLogReport(LogReport report);
 void logFinalReport(int killCount);
 void logParentInit();
-void logServerInfo(int port);
+void logServerInfo(uint16_t port);
 void logProcessMonitoringInit(char* processName, pid_t pid);
 void logProcessKill(pid_t pid, const char* name, unsigned long int duration);
 void logSelfDying(pid_t pid, const char* name, unsigned long int duration);
